added topbar to all pages
each page has the topbar in it now.

The index page is the only page that has the % completion. (also updated main.js to handle multiple locations of total progress percent, should we switch this to a class?).

Speech status/spectator status will be added into the topbar via their own divs.