Fixed checklist issue
It turns out specifically renaming the "40 Random Gates" subsection in misc.json will break the entire checklist functionality, so I just had to switch the name of this one back to get it working again. This behavior may be worth fixing later but I just used a note to explain the random gates instead.