reset and reload map page on load.
I couldn't find out why, but it seems that the page keeps the old data for calculate progress and doesn't update it until reload.

My solution is to reload the page after resetting location data.

There's probably a way to do it without a reload, but I do not know at this point in time.
closes #222 if this is good enough.