Updated Nirnroot descriptions
The auto generated instructions obsoleted many of the previous descriptions, and several were still referencing directions from the previous TSP route. All have been updated to be accurate for the new Nirnroute TSP.

Also had to put in a special case to not auto generate instructions for the Swamp Start because they were conflicting with the setup to align with the Ayleid Well.