fixed settings issue for new users
The iframe wasn't loading into the side panel on first load in from an incognito window, only the second time onward. Turns out the window.settings was being set to null by the loadCookie function, so I made it an async function that bottle necks the loading of the iframe until it has loaded correctly. This appears to have no noticeable effect on performance.

Small note: The very first iteration, window.settings is labeled as Object in the console, but it functions identically and all the cookie info is the exact same.