added auto width optimizations
On page load in, the divider position is set depending on what page the user is one. Everything is in place to allow for checking window dimensions, existing settings, etc.

TODO: The main guide page has something on it that is breaking the feature, but it can still be observed on the nirnroute page in the meantime.