updated popout layout
Now designed to work as an OBS browser source instead of chroma keyed.

- Background of page is transparent.
- Added a font from the game.
- Made the popout by two columns for now.

Additionally, made wayshrines worth 0 and moved the remaining points to the greater powers section since wayshrines are not tracked by save file drag/drop and are not in themselves specifically a requirement.

In the future, I plan on adding settings for the following:
- Width/height
- single column, double column