Set up a better formatted Archive section
This may eventually get pushed to a wiki page instead for ease of editing the text, but for now we have a more readable system for showing the user new incremental changes to the route logic.

Also added a link to the patch notes (aka archive) at the top of the actual guide.